room_booking_system/requirements.txt
asgiref==3.8.1
Babel==2.15.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-appointment==3.5.1
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0     
django-picklefield==3.2
django-q2==1.6.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
phonenumbers==8.13.37
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1